[feature] hw_L65: created 'user.routes.js', replaced user routes from 'server.js' to 'user.routes.js', changes applied
